// packages/types/src/interfaces.ts
export type DayType = "MONDAY" | "TUESDAY" | "WEDNESDAY" | "THURSDAY" | "FRIDAY" | "SATURDAY" | "SUNDAY";

export enum BoostTypeEnum {
  CHANCE = "CHANCE",
  EXPERIENCE = "EXPERIENCE",
  SHINY = "SHINY"
}
export type BoostType = "CHANCE" | "EXPERIENCE" | "SHINY";

export enum ItemTypeEnum {
  NONE = "NONE",
  BOOSTER = "BOOSTER",
  TROPHY = "TROPHY",
  PAINT_BUCKET = "PAINT_BUCKET",
  DAILY_SPINNY_WHEEL_TICKET = "DAILY_SPINNY_WHEEL_TICKET"
}
export type ItemType = "NONE" | "BOOSTER" | "TROPHY" | "PAINT_BUCKET" | "DAILY_SPINNY_WHEEL_TICKET";

export enum ItemShopItemTypeEnum {
  ITEM = "ITEM",
  BLOOK = "BLOOK",
  TITLE = "TITLE",
  BANNER = "BANNER",
  FONT = "FONT"
}
export type ItemShopItemType = "ITEM" | "BLOOK" | "TITLE" | "BANNER" | "FONT";

export enum ProductTypeEnum {
  ITEM = "ITEM",
  BLOOK = "BLOOK",
  TITLE = "TITLE",
  FONT = "FONT",
  BANNER = "BANNER",
  OTHER = "OTHER"
}
export type ProductType = "ITEM" | "BLOOK" | "TITLE" | "FONT" | "BANNER" | "OTHER";

export enum RarityAnimationTypeEnum {
  COMMON = "COMMON",
  RARE = "RARE",
  EPIC = "EPIC",
  LEGENDARY = "LEGENDARY",
  CHROMA = "CHROMA",
  MYTHICAL = "MYTHICAL"
}
export type RarityAnimationType = "COMMON" | "RARE" | "EPIC" | "LEGENDARY" | "CHROMA" | "MYTHICAL";
// ...
export type SpinnyWheelRewardType = "ITEM" | "BLOOK" | "TITLE" | "FONT" | "BANNER" | "TOKENS" | "CRYSTALS" | "DIAMONDS";
// ...
export type BlookObtainMethod = "UNKNOWN" | "PACK_OPEN" | "PURCHASE" | "STAFF";
// ...
export interface Auction {
  id: number;
  type: AuctionType;
  price: number;
  buyItNow: boolean;
  expiresAt: Date;
  createdAt: Date;
  updatedAt: Date;
  blookId: number | null;
  blook?: UserBlook | null;
  itemId: number | null;
  item?: UserItem | null;
  bids?: AuctionBid[];
  sellerId: string;
  seller?: User;
  buyerId: string | null;
  buyer?: User | null;
  delistedAt: Date | null;
}
// ...
export interface Banner {
  id: number;
  name: string;
  imageId: number;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
  userBanner?: UserBanner[];
  itemShop?: ItemShop[];
  products?: Product[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}

export interface Blacklist {
  id: number;
  ipAddressId: number;
  punishmentId: number;
  createdAt: Date;
  updatedAt: Date;
  ipAddress?: IpAddress;
  punishment?: Punishment;
}

export interface Blook {
  id: number;
  name: string;
  description: string | null;
  chance: number;
  price: number;
  rarityId: number;
  imageId: number;
  backgroundId: number;
  videoId: number | null;
  packId: number | null;
  onlyOnDay: DayType | null;
  isBig: boolean;
  shinyHue: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  background?: Resource;
  video?: Resource | null;
  image?: Resource;
  pack?: Pack | null;
  rarity?: Rarity;
  itemShops?: ItemShop[];
  products?: Product[];
  userBlooks?: UserBlook[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}

export interface Boost {
  id: number;
  userId: string;
  multiplier: number;
  type: BoostType;
  solo: boolean;
  createdAt: Date;
  updatedAt: Date;
  expiresAt: Date;
  user?: User;
}

export interface Emoji {
  id: number;
  name: string;
  imageId: number;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
}

export interface Font {
  id: number;
  name: string;
  resourceId: number;
  default: boolean;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
  itemShop?: ItemShop[];
  products?: Product[];
  users?: UserFont[];
  usersUsing?: User[];
  spinnyWheelRewards?: SpinnyWheelReward[];
  titles?: Title[];
}

export interface Group {
  id: number;
  name: string;
  description: string | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  permissions: PermissionType[];
  imageId: number | null;
  image?: Resource | null;
  products?: Product[];
  users?: UserGroup[];
  discordRoleId: string | null;
}

export interface Guild {
  id: number;
  name: string;
  description: string;
  color: string;
  experience: number;
  imageId: number | null;
  acceptingRequests: boolean;
  createdAt: Date;
  updatedAt: Date;
  image?: Resource | null;
  members?: UserGuild[];
  joinRequests?: UserGuildRequest[];
}

export interface IpAddress {
  id: number;
  ipAddress: string;
  createdAt: Date;
  updatedAt: Date;
  users?: User[];
  userIpAddresses?: UserIpAddress[];
  userSubscriptions?: UserSubscription[];
  blacklists?: Blacklist[];
  transactions?: Transaction[];
}

export interface Item {
  id: number;
  name: string;
  description: string | null;
  rarityId: number;
  imageId: number;
  type: ItemType;
  canUse: boolean;
  canAuction: boolean;
  canTrade: boolean;
  maxUses: number;
  boosterDuration: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  resource?: Resource;
  rarity?: Rarity;
  itemShops?: ItemShop[];
  products?: Product[];
  users?: UserItem[];
  spinnyWheels?: SpinnyWheel[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}

export interface ItemShop {
  id: number;
  type: ItemShopItemType;
  itemId: number | null;
  blookId: number | null;
  titleId: number | null;
  bannerId: number | null;
  fontId: number | null;
  price: number;
  enabled: boolean;
  weekly: boolean;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook | null;
  item?: Item | null;
  title?: Title | null;
  banner?: Banner | null;
  font?: Font | null;
}
// ...
export interface Pack {
  id: number;
  name: string;
  price: number;
  enabled: boolean;
  hidden: boolean;
  imageId: number;
  iconId: number;
  backgroundId: number;
  ambienceId: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook[];
  image?: Resource;
  icon?: Resource;
  background?: Resource;
  ambience?: Resource | null;
}

export interface Product {
  id: number;
  stripeProductId: string | null;
  stripePriceId: string | null;
  name: string;
  description: string | null;
  price: number;
  subscriptionPrice: number | null;
  imageId: number;
  color1: string;
  color2: string;
  type: ProductType;
  itemId: number | null;
  blookId: number | null;
  titleId: number | null;
  fontId: number | null;
  bannerId: number | null;
  groupId: number | null;
  tokens: number | null;
  diamonds: number | null;
  crystals: number | null;
  isSubscription: boolean | null;
  isQuantityCapped: boolean | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  image?: Resource;
  blook?: Blook | null;
  item?: Item | null;
  title?: Title | null;
  font?: Font | null;
  banner?: Banner | null;
  group?: Group | null;
  subscribers?: UserSubscription[];
  stores?: Store[];
  transactions?: Transaction[];
}

export interface Punishment {
  id: number;
  userId: string;
  type: PunishmentType;
  reason: string;
  expiresAt: Date;
  staffId: string;
  createdAt: Date;
  updatedAt: Date;
  staff?: User;
  user?: User;
  blacklists?: Blacklist[];
}

export interface Rarity {
  id: number;
  name: string;
  color: string;
  animationType: RarityAnimationType;
  experience: number;
  affectedByBooster: boolean;
  imageId: number | null;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook[];
  item?: Item[];
  image?: Resource | null;
}

export interface Resource {
  id: number;
  path: string;
  reference: string | null;
  createdAt: Date;
  updatedAt: Date;
  banners?: Banner[];
  blookBackgrounds?: Blook[];
  blooks?: Blook[];
  blookVideos?: Blook[];
  emojis?: Emoji[];
  fonts?: Font[];
  groups?: Group[];
  guilds?: Guild[];
  items?: Item[];
  packIcons?: Pack[];
  packImages?: Pack[];
  packBackgrounds?: Pack[];
  packAmbiences?: Pack[];
  products?: Product[];
  userBanners?: User[];
  rarities?: Rarity[];
}

export interface Room {
  id: number;
  name: string;
  public: boolean;
  createdAt: Date;
  updatedAt: Date;
  messages?: Message[];
  users?: User[];
}

export interface Session {
  id: string;
  userId: string;
  createdAt: Date;
  user?: User;
}

export interface SpinnyWheel {
  id: number;
  name: string;
  itemId: number;
  item?: Item;
  rewards?: SpinnyWheelReward[];
  createdAt: Date;
  updatedAt: Date;
}

export interface SpinnyWheelReward {
  id: number;
  spinnyWheelId: number;
  spinnyWheel?: SpinnyWheel;
  type: SpinnyWheelRewardType;
  chance: number;
  quantity: number;
  itemId: number | null;
  blookId: number | null;
  titleId: number | null;
  fontId: number | null;
  bannerId: number | null;
  tokens: number | null;
  crystals: number | null;
  diamonds: number | null;
  createdAt: Date;
  updatedAt: Date;
  item?: Item | null;
  blook?: Blook | null;
  title?: Title | null;
  font?: Font | null;
  banner?: Banner | null;
}

export interface Store {
  id: number;
  name: string;
  description: string;
  priority: number;
  products?: Product[];
  createdAt: Date;
  updatedAt: Date;
  active: boolean;
}

export interface Title {
  id: number;
  name: string;
  color: string;
  fontId: number;
  bold: boolean;
  italic: boolean;
  underline: boolean;
  priority: number;
  createdAt: Date;
  updatedAt: Date;
  font?: Font;
  itemShop?: ItemShop[];
  products?: Product[];
  users?: UserTitle[];
  usersUsing?: User[];
  spinnyWheelRewards?: SpinnyWheelReward[];
}

export interface Transaction {
  id: string;
  userId: string;
  amount: number;
  quantity: number;
  currency: CurrencyType;
  stripePaymentId: string | null;
  productId: number;
  paymentMethodId: number;
  ipAddressId: number;
  status: TransactionStatus;
  createdAt: Date;
  user?: User;
  product?: Product;
  paymentMethod?: UserPaymentMethod;
  ipAddress?: IpAddress;
}

export interface Upload {
  id: number;
  uploadId: string;
  filename: string;
  userId: string;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
  customAvatars?: User[];
  customBanners?: User[];
  newsPosts?: NewsPost[];
}

export interface User {
  id: string;
  username: string;
  password: string;
  email: string | null;
  emailVerified: boolean;
  avatarId: number | null;
  customAvatarId: number | null;
  bannerId: number | null;
  customBannerId: number | null;
  titleId: number;
  fontId: number;
  color: string;
  tokens: number;
  diamonds: number;
  crystals: number;
  experience: number;
  lastClaimed: Date;
  ipAddressId: number;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  lastSeen: Date;
  readRulesAt: Date | null;
  stripeCustomerId: string | null;
  isSystem: boolean;
  isAi: boolean;
  subscriptions?: UserSubscription[];
  permissions: PermissionType[];
  bids?: AuctionBid[];
  messages?: Message[];
  rooms?: Room[];
  sessions?: Session[];
  friendedBy?: User[];
  friends?: User[];
  blockedBy?: User[];
  blocked?: User[];
  avatar?: UserBlook | null;
  banner?: Resource | null;
  customAvatar?: Upload | null;
  customBanner?: Upload | null;
  font?: Font;
  title?: Title;
  banners?: UserBanner[];
  initiallyObtainedBlooks?: UserBlook[];
  blooks?: UserBlook[];
  discord?: UserDiscord | null;
  groups?: UserGroup[];
  guild?: UserGuild[];
  guildRequests?: UserGuildRequest[];
  ipAddress?: IpAddress;
  ipAddresses?: UserIpAddress[];
  initiallyObtainedItems?: UserItem[];
  items?: UserItem[];
  OAuth?: UserOAuth[];
  authMethods?: UserWebAuthn[];
  paymentMethods?: UserPaymentMethod[];
  givenPunishments?: Punishment[];
  punishments?: Punishment[];
  settings?: UserSetting | null;
  statistics?: UserStatistic | null;
  titles?: UserTitle[];
  fonts?: UserFont[];
  auctions?: Auction[];
  wonAuctions?: Auction[];
  boosts?: Boost[];
  uploads?: Upload[];
  newsPosts?: NewsPost[];
  votedNewsPosts?: UserNewsPostVote[];
  transactions?: Transaction[];
  auditLogs?: AuditLog[];
  undoneLogs?: AuditLog[];
}

export interface UserBanner {
  id: number;
  userId: string;
  bannerId: number;
  createdAt: Date;
  updatedAt: Date;
  banner?: Banner;
  user?: User;
}

export interface UserBlook {
  id: number;
  userId: string;
  blookId: number;
  sold: boolean;
  shiny: boolean;
  initialObtainerId: string;
  obtainedBy: BlookObtainMethod;
  serial: number | null;
  canSell: boolean;
  canTrade: boolean;
  canAuction: boolean;
  createdAt: Date;
  updatedAt: Date;
  blook?: Blook;
  initialObtainer?: User;
  user?: User;
  auctions?: Auction[];
  avatars?: User[];
}
// ...
export interface UserItem {
  id: number;
  userId: string;
  itemId: number;
  usesLeft: number;
  initialObtainerId: string;
  obtainedBy: ItemObtainMethod;
  createdAt: Date;
  updatedAt: Date;
  initialObtainer?: User;
  item?: Item;
  user?: User;
  auctions?: Auction[];
}
// ...
export interface UserSubscription {
  id: number;
  userId: string;
  stripeSubscriptionId: string;
  productId: number;
  status: UserSubscriptionStatus;
  ipAddressId: number;
  createdAt: Date;
  updatedAt: Date;
  expiresAt: Date | null;
  user?: User;
  product?: Product;
  ipAddress?: IpAddress;
}